# Optimize the size of Instruction