--- a/jit/src/instructions.rs
+++ b/jit/src/instructions.rs
@@ -320,7 +320,7 @@ impl<'a, 'b> FunctionCompiler<'a, 'b> {
                     _ => Err(JitCompileError::NotSupported),
                 }
             }
-            Instruction::BinaryOperation { op, .. } => {
+            Instruction::BinaryOperation { op } | Instruction::BinaryOperationInplace { op } => {
                 // the rhs is popped off first
                 let b = self.stack.pop().ok_or(JitCompileError::BadBytecode)?;
                 let a = self.stack.pop().ok_or(JitCompileError::BadBytecode)?;