This reduces the size of the Instruction enum from 56 bytes to 16 bytes, which should help with memory usage as well as a (very) slight performance boost:
```
Benchmark #1: ./rustpython-norm benchmarks/pystone.py 50000
  Time (mean ± σ):      3.655 s ±  0.119 s    [User: 3.637 s, System: 0.007 s]
  Range (min … max):    3.522 s …  3.853 s    10 runs
 
Benchmark #2: ./rustpython-opt benchmarks/pystone.py 50000
  Time (mean ± σ):      3.624 s ±  0.074 s    [User: 3.601 s, System: 0.008 s]
  Range (min … max):    3.557 s …  3.783 s    10 runs
 
Summary
  './rustpython-opt benchmarks/pystone.py 50000' ran
    1.01 ± 0.04 times faster than './rustpython-norm benchmarks/pystone.py 50000'
```

The instructions I changed were either a) flattening an inner enum, like `CallFunction { CallType } -> CallFunction{Positional,Keyword,Ex}` b) not in many hot paths, like Import, or c) provided redundant information, like SetupLoop, whose `start` field always pointed to `lasti+1`

Edit: well, SetupLoop isn't actually like that, but there was a way around it