@coolreader18 2020-12-10
Here's the benchmark for the lz4_flex change:
```
Benchmark #1: ./rustpython-norm -c ""
  Time (mean ± σ):      53.1 ms ±  13.2 ms    [User: 45.4 ms, System: 7.3 ms]
  Range (min … max):    36.3 ms …  77.2 ms    66 runs
 
Benchmark #2: ./rustpython-opt -c ""
  Time (mean ± σ):      15.2 ms ±   7.4 ms    [User: 11.0 ms, System: 4.2 ms]
  Range (min … max):     7.7 ms …  33.8 ms    206 runs
 
Summary
  './rustpython-opt -c ""' ran
    3.50 ± 1.92 times faster than './rustpython-norm -c ""'
```